odoo_project/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astro-sdk-python[snowflake] >= 1.1.0
pandas
pyyaml